feat: Add support for playoff-bracket and playoff-series endpoints
- Updated NHL franchise, team, player, and game SQL migrations to use INTEGER instead of SMALLINT for IDs and foreign keys for better scalability.
- Added a new migration for nhl_playoff_series table to store playoff series data.
- Implemented get_nhl_playoff_series method in NhlWebApi to fetch and cache playoff series data from the API.
- Added NhlPlayoffSeries and NhlPlayoffBracket models to represent playoff series data.
- Updated serde_helpers for improved JSON handling and logging.
- Adjusted logging levels from warn to debug for non-critical issues to reduce log noise.

+++ migrations/03_create_nhl_franchise.sql
@@ -1,9 +1,9 @@
 CREATE TABLE IF NOT EXISTS nhl_franchise (
-    id SMALLINT PRIMARY KEY,
+    id INTEGER PRIMARY KEY,
     full_name TEXT NOT NULL,
     team_common_name TEXT NOT NULL,
     team_place_name TEXT NOT NULL,
     raw_json JSONB NOT NULL,
-    api_cache_endpoint TEXT NOT NULL REFERENCES api_cache(endpoint),
+    api_cache_endpoint TEXT NOT NULL REFERENCES api_cache (endpoint),
     last_updated TIMESTAMP DEFAULT now()
 )
